refactor: create network utils file
